requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==3.2.13
djangorestframework==3.13.1
flake8==4.0.1
idna==3.3
importlib-metadata==4.2.0
mccabe==0.6.1
pycodestyle==2.8.0
pyflakes==2.4.0
python-dotenv==0.20.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
stripe==3.1.0
typing_extensions==4.2.0
urllib3==1.26.9
zipp==3.8.0
